Update FileReader/FileWriter test cases to match new FileWriter implementation.
In 0.9.4 and earlier versions, FileWriter could be obtained synchronously using it's constructor: function FileWriter(filePath, append).  This constructor is a deviation from the W3C File API: Writer spec.  It is also not possible to obtain a FileWriter synchronously on iOS.

As of 0.9.5, the FileWriter constructor requires a File object.  It is also possible to create a FileWriter using the asynchronous function: FileEntry.createWriter(successCallback, errorCallback).  FileEntry is part of the File API: Directories and System W3C spec, which was added in 0.9.5.