Add W3C File: Directories and System test cases.
The current spec supports PERSISTENT and TEMPORARY file system types.
Both types are retrieved at the beginning of the file tests, but only
one is used for the tests.  It is recommended that the file tests be
repeated for each type.  To change types, edit the return value of
getFileSystemRoot function in file.tests.js.