CB-6121 (Windows command line script for MS)
Windows command line script for Mobile spec tests.

Description:
Basically is a command line script file (With a little of VBS), that
uses the Cordova Command Line Interface, git for windows, and node
cordova-cli module to create and assist the test process, using cordova
mobile specs test suite, that includes manual and automated tests, to
test plugins capabilities, as well as stability through versions.
The script capabilities helps to simplify the test process, obtaining
all resources required or giving the chance to set them in place, or set
where they are.
This script file it uses the cordova CLI capabilities to create, add
platforms/plugins, as well as build, emulate and deploy to phone, it
also uses the coho script to update some repositories, and get several
resources if it's required.

Support:
Tested under: Windows 7, Windows 8.
Not Tested: Windows XP, Vista, and Server.
Phone platforms supported: Android, Blackberry10, Windows Phone 8,
Windows 8 [Not fully supported, just creation not build and deploy].

Capabilities:
-Creates cordova based project for mobile spec.
-Add cordova supported platforms (Android, Blackberry10, Windows Phone
8, and -Windows 8[Not fully supported]
-Add option for the whole set of plugins from cordova-coho repository
(Local).
-Add option for custom, or personal single plugins to add to the
project, just providing the location path for it.
-It adds every required file and configuration from the mobile spec
repository, as well as the tests to the www folder.
-It provides and transfers an updated JavaScript file for each platform,
obtained through the use of the grunt task.
-Deploy to emulator, it provide the option to deploy into a platform
emulator.
-Deploy to phone, it also provide the option to deploy into a physical
phone..
-It Gets cordova-coho repository if is not present, letting you chose
where do you want to download it.
-It allows to get cordova-coho repository update, using the coho script.
-It allows to easily switch between Menus, go to add plugins, platforms,
deploy to emulator or deploy to phone.

Requirements:
Git for windows [http://git-scm.com/download/win]
Node.js [http://nodejs.org/],
Cordova-CLI module for node.js installed.

Usage:
Just double clic, and follow the program options, it's really easy to
use.

Note: It uses default cordova commands to add plugins, but it can be
change to use plugman (If it's installed), at line 320 it has added the
line to use plugman, just uncomment it, and comment the previous one to
use it.