[CB-6822] plugman compatible with windows
When createmobilespec module is used and the --plugman argument is
present, it will determine if the development platform is windows to add
node as part of the command.

process.argv[0] it will retrieve the node from the current process, and
it will cover the scenario where the nodejs installation it has a custom
path or it's not present in system PATH environmental variable.

If:
/path/to/nodejs/node createmobilespec.js [args]
process.argv[0] it will retrieve: /path/to/nodejs/node
to invoke plugman main.js file.