CB-7380 update tests for file URL
It was previously checking that a file URL would fail, but file URLs should
always be accepted otherwise local loading into the webview won't work. So
the file URL test was broken into 2 separate tests. The first was changed to
expect success on an empty whitelist against a file URL, and the pattern match
part of the original file URL test was put in a second (new) test to use
an HTTP URL.