Tip to delete existing mobilespec if it exists (#178)
before running `createmobilespec.js`

(this seems to help me avoid the quirks I documented before)

Note that this proposal is expected to conflict with PR #154, which I did not get a chance to review yet. I am proposing this tip anyway due to its importance IMHO, to capture the info and avoid losing track of it. I hope to review the documentation updates proposed in PR #154 within the next few days.